# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Energy Forecast API
  description: API for energy demand forecasting and renewable energy optimization
  version: 1.0.0
  contact:
    name: Energy Forecast Team
    
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.energyforecast.com
    description: Production server

security:
  - ApiKeyAuth: []
  - BearerAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      
  schemas:
    WeatherData:
      type: object
      properties:
        temperature:
          type: number
          description: Temperature in Celsius
          minimum: -50
          maximum: 50
        humidity:
          type: number
          description: Relative humidity percentage
          minimum: 0
          maximum: 100
        cloud_cover:
          type: number
          description: Cloud cover percentage
          minimum: 0
          maximum: 100
          
    ForecastRequest:
      type: object
      required:
        - city
        - start_date
        - end_date
      properties:
        city:
          type: string
          description: City name
          enum: [Mumbai, Delhi]
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        weather_forecast:
          type: array
          items:
            $ref: '#/components/schemas/WeatherData'
            
    ForecastResponse:
      type: object
      properties:
        demand_forecast:
          type: object
          additionalProperties:
            type: number
        energy_mix:
          type: object
          additionalProperties:
            type: object
        confidence_intervals:
          type: object
          properties:
            lower:
              type: object
            upper:
              type: object

paths:
  /health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: Service health status
          
  /forecast/demand:
    post:
      summary: Generate energy demand forecast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForecastRequest'
      responses:
        '200':
          description: Forecast generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded
          
  /forecast/cities:
    get:
      summary: List available cities
      responses:
        '200':
          description: List of supported cities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  
  /data/validate:
    post:
      summary: Validate input data
      parameters:
        - name: city
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeatherData'
      responses:
        '200':
          description: Validation results